integrate with coretify & cleanup code

+++ src/components/project.jsx
@@ -4,7 +4,7 @@ import { gql, useQuery } from '@apollo/client';
 import { primaryFontColor, ternaryColor } from '@/app/theme/globalTheme';
 import {
     Heading, Tabs, TabList, TabPanels, Tab, TabPanel, useToast,
-    Box, ListItem, Text, Input, Button, OrderedList, Flex,
+    Box, Text, Input, Button, Flex,
     TableContainer,
     Table,
     Thead,
@@ -27,7 +27,7 @@ const GET_PROJECT_DATA = gql`
     }`;
 
 
-function Project({ onOpenModalAuth }) {
+function Project() {
 
     const { loading, error, data, refetch } = useQuery(GET_PROJECT_DATA);
 
@@ -44,30 +44,7 @@ function Project({ onOpenModalAuth }) {
     const [imageUrl, setImageUrl] = useState("");
     const [skillsUrl, setSkillsUrl] = useState("");
 
-    const handleAuthInfo = () => {
-        if (sessionStorage.getItem("basic_auth")) {
-            const authInfo = JSON.parse(sessionStorage.getItem("basic_auth"));
-            if (authInfo.is_valid == true && authInfo.username && authInfo.password) {
-                return [authInfo.username, authInfo.password, authInfo.is_valid];
-            } else return ['', '', false];
-        } else return ['', '', false];
-    }
-
-    const handleAuthAction = (func) => {
-        if (sessionStorage.getItem('basic_auth')) {
-            const authInfo = JSON.parse(sessionStorage.getItem('basic_auth'));
-            if (authInfo.is_valid == true) func();
-            else onOpenModalAuth();
-        } else onOpenModalAuth()
-    }
-
     const cleanStateOperation = (toastMessage, toastStatus) => {
-        const prevAuthInfo = JSON.parse(sessionStorage.getItem('basic_auth'));
-        if (toastStatus === "success") {
-            sessionStorage.setItem('basic_auth', JSON.stringify({ ...prevAuthInfo, is_valid: true }));
-        } else if (toastStatus === "error") {
-            sessionStorage.setItem('basic_auth', JSON.stringify({ ...prevAuthInfo, is_valid: false }));
-        }
         toast({
             title: toastMessage,
             status: toastStatus,
@@ -89,23 +66,28 @@ function Project({ onOpenModalAuth }) {
         try {
             setIsLoading(true);
 
-            const [username, password] = handleAuthInfo();
+            const token = localStorage.getItem('token');
 
-            const encodedCredentials = btoa(`${username}:${password}`);
+            // decline action if token not exist.
+            if (!token) {
+                cleanStateOperation('Failed to create data.', 'error');
+                return;
+            }
 
             const response = await fetch('/api/projects', {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
-                    'Authorization': `Basic ${encodedCredentials}`
+                    'Authorization': token
                 },
                 body: JSON.stringify({ heading, text, imageUrl, skillsUrl })
             });
-            if (!response.ok) {
-                cleanStateOperation('Failed to Add data.', 'error');
+
+            if (response.status === 201) {
+                cleanStateOperation('201: Successfully created data.', 'success');
                 return;
             }
-            cleanStateOperation('200: Successfully Add data.', 'success');
+            else cleanStateOperation('Failed to create data.', 'error');
         } catch (error) {
             console.error(error.message);
             cleanStateOperation('Failed to Add data.', 'error');
@@ -116,20 +98,26 @@ function Project({ onOpenModalAuth }) {
         try {
             setIsLoading(true);
 
-            const [username, password] = handleAuthInfo();
+            const token = localStorage.getItem('token');
 
-            const encodedCredentials = btoa(`${username}:${password}`);
+            // decline action if token not exist.
+            if (!token) {
+                cleanStateOperation('Failed to create data.', 'error');
+                return;
+            }
 
             const response = await fetch(`/api/projects/${id}`, {
                 method: 'DELETE',
                 headers: {
-                    'Authorization': `Basic ${encodedCredentials}`,
+                    'Authorization': token,
                 },
             });
-            if (!response.ok) {
-                cleanStateOperation('Failed to delete data.', 'error');
+
+            if (response.status === 204) {
+                cleanStateOperation('204: Successfully delete data.', 'success');
+                return;
             }
-            cleanStateOperation('200: Successfully Delete data.', 'success');
+            else cleanStateOperation('Failed to delete data.', 'error');
         } catch (error) {
             console.error(error.message);
             cleanStateOperation('Failed to delete data.', 'error');
@@ -180,7 +168,7 @@ function Project({ onOpenModalAuth }) {
                                                             <Text>{object.text}</Text>
                                                         </Td>
                                                         <Td>
-                                                            <Button onClick={() => handleAuthAction(() => handleDeleteData(object.id))}
+                                                            <Button onClick={() => handleDeleteData(object.id)}
                                                                 isLoading={isLoading}
                                                                 ml={2} variant={'outline'} colorScheme={'red'} size={'sm'}>
                                                                 Delete
@@ -225,7 +213,7 @@ function Project({ onOpenModalAuth }) {
                                         colorScheme='purple' borderColor={"#536189"} focusBorderColor={ternaryColor}
                                         type="text"
                                     />
-                                    <Button onClick={() => handleAuthAction(() => handlePostData())} isLoading={isLoading} my={3} fontWeight={'bold'} colorScheme='purple' color={'black'}>Add +</Button>
+                                    <Button onClick={() => handlePostData()} isLoading={isLoading} my={3} fontWeight={'bold'} colorScheme='purple' color={'black'}>Add +</Button>
                                 </TabPanel>
                             </TabPanels>
                         </Tabs>
@@ -274,7 +262,7 @@ function Project({ onOpenModalAuth }) {
                                                             <Text>{object.text}</Text>
                                                         </Td>
                                                         <Td>
-                                                            <Button onClick={() => handleAuthAction(() => handleDeleteData(object.id))}
+                                                            <Button onClick={() => handleDeleteData(object.id)}
                                                                 isLoading={isLoading}
                                                                 ml={2} variant={'outline'} colorScheme={'red'} size={'sm'}>
                                                                 Delete
@@ -319,7 +307,7 @@ function Project({ onOpenModalAuth }) {
                                         colorScheme='purple' borderColor={"#536189"} focusBorderColor={ternaryColor}
                                         type="text"
                                     />
-                                    <Button onClick={() => handleAuthAction(() => handlePostData())} isLoading={isLoading} my={3} fontWeight={'bold'} colorScheme='purple' color={'black'}>Add +</Button>
+                                    <Button onClick={() => handlePostData()} isLoading={isLoading} my={3} fontWeight={'bold'} colorScheme='purple' color={'black'}>Add +</Button>
                                 </TabPanel>
                             </TabPanels>
                         </Tabs>
